🐛 Fixe le problème de build avec l'export de authOptions dans le fichier de route Next-Auth
