+++ src/app/api/auth/[...nextauth]/route.ts
@@ -1,79 +1,24 @@
-import NextAuth, { AuthOptions } from "next-auth";
-import CredentialsProvider from "next-auth/providers/credentials";
-import { connectDB } from "@/lib/db";
-import { User } from "@/models/user";
-import bcrypt from "bcrypt";
+// src/app/api/auth/[...nextauth]/route.ts
+/**
+ * Route d'authentification NextAuth
+ * @module api/auth/route
+ *
+ * Cette route gère les requêtes d'authentification via NextAuth.
+ * Elle utilise la configuration définie dans options.ts pour gérer
+ * l'authentification et les sessions utilisateur.
+ */
 
-export const authOptions: AuthOptions = {
-  providers: [
-    CredentialsProvider({
-      name: "credentials",
-      credentials: {
-        email: { label: "Email", type: "text" },
-        password: { label: "Password", type: "password" },
-      },
-      async authorize(credentials) {
-        try {
-          if (!credentials?.email || !credentials?.password) {
-            return null;
-          }
+import NextAuth from "next-auth";
+import { authOptions } from "./options";
 
-          await connectDB();
-
-          const user = await User.findOne({ email: credentials.email });
-          if (!user) {
-            return null;
-          }
-
-          const isPasswordValid = await bcrypt.compare(
-            credentials.password,
-            user.password
-          );
-
-          if (!isPasswordValid) {
-            return null;
-          }
-
-          return {
-            id: user._id.toString(),
-            name: null,
-            email: user.email,
-            image: null,
-          };
-        } catch (error) {
-          console.error("Erreur d'authentification:", error);
-          return null;
-        }
-      },
-    }),
-  ],
-  session: {
-    strategy: "jwt",
-  },
-  pages: {
-    signIn: "/auth/connexion",
-  },
-  callbacks: {
-    async jwt({ token, user }) {
-      if (user) {
-        token.id = user.id;
-        token.name = user.name;
-        token.email = user.email;
-        token.picture = user.image;
-      }
-      return token;
-    },
-    async session({ session, token }) {
-      if (session.user) {
-        session.user.id = token.id;
-        session.user.name = token.name;
-        session.user.email = token.email;
-        session.user.image = token.picture;
-      }
-      return session;
-    },
-  },
-};
-
+/**
+ * Handler NextAuth pour gérer les requêtes GET et POST
+ * Utilise les options de configuration définies dans authOptions
+ */
 const handler = NextAuth(authOptions);
+
+/**
+ * Export des handlers pour les méthodes GET et POST
+ * Nécessaire pour le fonctionnement des routes API dans Next.js 13+
+ */
 export { handler as GET, handler as POST };
